feat(js): remove ts implementation of ts processing
